fix declaration incompatibility

+++ implementation/lfo.h
@@ -3,7 +3,7 @@
 
 #include <ezdsp5535.h>
 
-void fRate(int r);
+void fRate(Int16 r);
 Int16 LFO();
 
 #endif /*LFO_H_*/
